+++ src/neo4j/_async/io/_pool.py
@@ -46,8 +46,8 @@ from ..._deadline import (
 from ..._exceptions import BoltError
 from ..._routing import RoutingTable
 from ...api import (
+    check_access_mode,
     READ_ACCESS,
-    WRITE_ACCESS,
 )
 from ...exceptions import (
     ClientError,
@@ -670,6 +670,7 @@ class AsyncBoltPool(AsyncIOPool):
     ):
         # The access_mode and database is not needed for a direct connection,
         # it's just there for consistency.
+        access_mode = check_access_mode(access_mode)
         _check_acquisition_timeout(timeout)
         log.debug(
             "[#0000]  _: <POOL> acquire direct connection, "
@@ -852,6 +853,8 @@ class AsyncNeo4jPool(AsyncIOPool):
                 "[#0000]  _: <POOL> failed to fetch routing info from %r",
                 address,
             )
+            # TODO: 7.0 - when Python 3.11+ is the minimum,
+            #       use exception groups instead of swallowing discovery errors
             return None
         else:
             servers = new_routing_info[0]["servers"]
@@ -1066,8 +1069,6 @@ class AsyncNeo4jPool(AsyncIOPool):
 
         :returns: `True` if an update was required, `False` otherwise.
         """
-        from ...api import READ_ACCESS
-
         async with self.refresh_lock:
             for database_ in list(self.routing_tables.keys()):
                 # Remove unused databases in the routing table
@@ -1114,8 +1115,6 @@ class AsyncNeo4jPool(AsyncIOPool):
 
     async def _select_address(self, *, access_mode, database):
         """Select the address with the fewest in-use connections."""
-        from ...api import READ_ACCESS
-
         async with self.refresh_lock:
             routing_table = self.routing_tables.get(database)
             if routing_table:
@@ -1152,14 +1151,8 @@ class AsyncNeo4jPool(AsyncIOPool):
         unprepared=False,
         database_callback=None,
     ):
-        if access_mode not in {WRITE_ACCESS, READ_ACCESS}:
-            # TODO: 6.0 - change this to be a ValueError
-            raise ClientError(f"Non valid 'access_mode'; {access_mode}")
-        _check_acquisition_timeout(timeout)
-
-        from ...api import check_access_mode
-
         access_mode = check_access_mode(access_mode)
+        _check_acquisition_timeout(timeout)
 
         target_database = database.name