+++ src/neo4j/_sync/io/_bolt4.py
@@ -19,10 +19,10 @@ from ssl import SSLSocket
 
 from ..._api import TelemetryAPI
 from ..._exceptions import BoltProtocolError
+from ..._io import BoltProtocolVersion
 from ...api import (
     READ_ACCESS,
     SYSTEM_DATABASE,
-    Version,
 )
 from ...exceptions import (
     ConfigurationError,
@@ -62,7 +62,7 @@ class Bolt4x0(Bolt):
     This is supported by Neo4j versions 4.0-4.4.
     """
 
-    PROTOCOL_VERSION = Version(4, 0)
+    PROTOCOL_VERSION = BoltProtocolVersion(4, 0)
 
     ssr_enabled = False
 
@@ -188,7 +188,7 @@ class Bolt4x0(Bolt):
         if imp_user is not None:
             raise ConfigurationError(
                 "Impersonation is not supported in Bolt Protocol "
-                f"{self.PROTOCOL_VERSION!r}. Trying to impersonate "
+                f"{self.PROTOCOL_VERSION}. Trying to impersonate "
                 f"{imp_user!r}."
             )
         dehydration_hooks, hydration_hooks = self._default_hydration_hooks(
@@ -247,7 +247,7 @@ class Bolt4x0(Bolt):
         if imp_user is not None:
             raise ConfigurationError(
                 "Impersonation is not supported in Bolt Protocol "
-                f"{self.PROTOCOL_VERSION!r}. Trying to impersonate "
+                f"{self.PROTOCOL_VERSION}. Trying to impersonate "
                 f"{imp_user!r}."
             )
         if (
@@ -359,7 +359,7 @@ class Bolt4x0(Bolt):
         if imp_user is not None:
             raise ConfigurationError(
                 "Impersonation is not supported in Bolt Protocol "
-                f"{self.PROTOCOL_VERSION!r}. Trying to impersonate "
+                f"{self.PROTOCOL_VERSION}. Trying to impersonate "
                 f"{imp_user!r}."
             )
         if (
@@ -529,7 +529,7 @@ class Bolt4x1(Bolt4x0):
     This is supported by Neo4j versions 4.1 - 4.4.
     """
 
-    PROTOCOL_VERSION = Version(4, 1)
+    PROTOCOL_VERSION = BoltProtocolVersion(4, 1)
 
     def get_base_headers(self):
         # Bolt 4.1 passes the routing context, originally taken from
@@ -551,7 +551,7 @@ class Bolt4x2(Bolt4x1):
     This is supported by Neo4j version 4.2 - 4.4.
     """
 
-    PROTOCOL_VERSION = Version(4, 2)
+    PROTOCOL_VERSION = BoltProtocolVersion(4, 2)
 
     SKIP_REGISTRATION = False
 
@@ -563,7 +563,7 @@ class Bolt4x3(Bolt4x2):
     This is supported by Neo4j version 4.3 - 4.4.
     """
 
-    PROTOCOL_VERSION = Version(4, 3)
+    PROTOCOL_VERSION = BoltProtocolVersion(4, 3)
 
     def get_base_headers(self):
         headers = super().get_base_headers()
@@ -581,7 +581,7 @@ class Bolt4x3(Bolt4x2):
         if imp_user is not None:
             raise ConfigurationError(
                 "Impersonation is not supported in Bolt Protocol "
-                f"{self.PROTOCOL_VERSION!r}. Trying to impersonate "
+                f"{self.PROTOCOL_VERSION}. Trying to impersonate "
                 f"{imp_user!r}."
             )
         dehydration_hooks, hydration_hooks = self._default_hydration_hooks(
@@ -668,7 +668,7 @@ class Bolt4x4(Bolt4x3):
     This is supported by Neo4j version 4.4.
     """
 
-    PROTOCOL_VERSION = Version(4, 4)
+    PROTOCOL_VERSION = BoltProtocolVersion(4, 4)
 
     def route(
         self,