+++ src/neo4j/_sync/io/_bolt5.py
@@ -23,11 +23,9 @@ from ..._api import TelemetryAPI
 from ..._async_compat.util import Util
 from ..._codec.hydration import v2 as hydration_v2
 from ..._exceptions import BoltProtocolError
+from ..._io import BoltProtocolVersion
 from ..._meta import BOLT_AGENT_DICT
-from ...api import (
-    READ_ACCESS,
-    Version,
-)
+from ...api import READ_ACCESS
 from ...exceptions import (
     DatabaseUnavailable,
     ForbiddenOnReadOnlyDatabase,
@@ -62,7 +60,7 @@ log = getLogger("neo4j.io")
 class Bolt5x0(Bolt):
     """Protocol handler for Bolt 5.0."""
 
-    PROTOCOL_VERSION = Version(5, 0)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 0)
 
     HYDRATION_HANDLER_CLS = hydration_v2.HydrationHandler
 
@@ -593,7 +591,7 @@ class ClientStateManager5x1(ClientStateManager):
 class Bolt5x1(Bolt5x0):
     """Protocol handler for Bolt 5.1."""
 
-    PROTOCOL_VERSION = Version(5, 1)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 1)
 
     supports_re_auth = True
 
@@ -684,7 +682,7 @@ class Bolt5x1(Bolt5x0):
 
 
 class Bolt5x2(Bolt5x1):
-    PROTOCOL_VERSION = Version(5, 2)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 2)
 
     supports_notification_filtering = True
 
@@ -869,7 +867,7 @@ class Bolt5x2(Bolt5x1):
 
 
 class Bolt5x3(Bolt5x2):
-    PROTOCOL_VERSION = Version(5, 3)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 3)
 
     def get_base_headers(self):
         headers = super().get_base_headers()
@@ -878,7 +876,7 @@ class Bolt5x3(Bolt5x2):
 
 
 class Bolt5x4(Bolt5x3):
-    PROTOCOL_VERSION = Version(5, 4)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 4)
 
     def telemetry(
         self,
@@ -907,7 +905,7 @@ class Bolt5x4(Bolt5x3):
 
 
 class Bolt5x5(Bolt5x4):
-    PROTOCOL_VERSION = Version(5, 5)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 5)
 
     def get_base_headers(self):
         headers = super().get_base_headers()
@@ -1107,7 +1105,7 @@ class Bolt5x5(Bolt5x4):
 
 
 class Bolt5x6(Bolt5x5):
-    PROTOCOL_VERSION = Version(5, 6)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 6)
 
     def _make_enrich_statuses_handler(self, wrapped_handler=None):
         def handler(metadata):
@@ -1139,7 +1137,7 @@ class Bolt5x6(Bolt5x5):
 
 
 class Bolt5x7(Bolt5x6):
-    PROTOCOL_VERSION = Version(5, 7)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 7)
 
     DEFAULT_ERROR_DIAGNOSTIC_RECORD = (
         Bolt5x5.DEFAULT_STATUS_DIAGNOSTIC_RECORD
@@ -1232,7 +1230,7 @@ class Bolt5x7(Bolt5x6):
 
 
 class Bolt5x8(Bolt5x7):
-    PROTOCOL_VERSION = Version(5, 8)
+    PROTOCOL_VERSION = BoltProtocolVersion(5, 8)
 
     @property
     def ssr_enabled(self) -> bool: